pyproject.toml
[tool.poetry]
name = "mlservice-tabular"
version = "0.1.0"
description = ""
authors = ["Bo Chen <bochen0909@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10"
pymlservice = "^0.1.1"
xgboost = "^2.1.3"
lightgbm = "^4.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
